@@ -8,31 +8,37 @@
             "name": "Rookie0"
         }
     ],
-    "require-dev": {
-        "nextcloud/coding-standard": "^v1.1.1",
-        "nextcloud/ocp": "dev-stable28",
-        "phpunit/phpunit": "^9",
-        "psalm/phar": "^5.17.0",
-        "sabre/dav": "^4.1",
-        "sabre/xml": "^2.2",
-        "symfony/event-dispatcher": "^5.3.11"
-    },
+    "require": {
+		"bamarni/composer-bin-plugin": "^1.8",
+		"php": "^8.1"
+	},
+	"require-dev": {
+		"nextcloud/ocp": "dev-stable29",
+		"roave/security-advisories": "dev-latest"
+	},
     "scripts": {
-        "lint": "find . -name \\*.php -not -path './vendor/*' -print0 | xargs -0 -n1 php -l",
-        "cs:check": "php-cs-fixer fix --dry-run --diff",
-        "cs:fix": "php-cs-fixer fix",
-        "psalm": "psalm.phar --threads=1",
-        "psalm:update-baseline": "psalm.phar --threads=1 --update-baseline",
-        "psalm:update-baseline:force": "psalm.phar --threads=1 --update-baseline --set-baseline=tests/psalm-baseline.xml",
-        "psalm:clear": "psalm.phar --clear-cache && psalm --clear-global-cache",
-        "psalm:fix": "psalm.phar --alter --issues=InvalidReturnType,InvalidNullableReturnType,MissingParamType,InvalidFalsableReturnType"
+		"post-install-cmd": [
+			"@composer bin all install --ansi"
+		],
+		"post-update-cmd": [
+			"@composer bin all update --ansi"
+		],
+		"lint": "find . -name \\*.php -not -path './vendor/*' -not -path './vendor-bin/*' -not -path './build/*' -print0 | xargs -0 -n1 php -l",
+		"cs:check": "php-cs-fixer fix --dry-run --diff",
+		"cs:fix": "php-cs-fixer fix",
+		"psalm": "psalm --threads=1 --no-cache",
+		"test:unit": "phpunit tests -c tests/phpunit.xml --colors=always --fail-on-warning --fail-on-risky",
+		"openapi": "generate-spec",
+		"rector": "rector && composer cs:fix"
     },
     "config": {
-        "allow-plugins": {
-            "composer/package-versions-deprecated": true
-        },
-        "platform": {
-            "php": "8.0"
-        }
-    }
+		"allow-plugins": {
+			"bamarni/composer-bin-plugin": true
+		},
+		"optimize-autoloader": true,
+		"sort-packages": true,
+		"platform": {
+			"php": "8.1"
+		}
+	}
 }
