@@ -1,10 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace OCA\SharingPath\Tests\Integration\Controller;
 
 use OCP\AppFramework\App;
 use Test\TestCase;
-
 
 /**
  * This test shows how to make a small Integration Test. Query your class
@@ -13,17 +14,17 @@ use Test\TestCase;
  */
 class AppTest extends TestCase {
 
-    private $container;
+	private $container;
 
-    public function setUp() {
-        parent::setUp();
-        $app = new App('sharingpath');
-        $this->container = $app->getContainer();
-    }
+	public function setUp(): void {
+		parent::setUp();
+		$app = new App('sharingpath');
+		$this->container = $app->getContainer();
+	}
 
-    public function testAppInstalled() {
-        $appManager = $this->container->query('OCP\App\IAppManager');
-        $this->assertTrue($appManager->isInstalled('sharingpath'));
-    }
+	public function testAppInstalled(): void {
+		$appManager = $this->container->query('OCP\App\IAppManager');
+		$this->assertTrue($appManager->isInstalled('sharingpath'));
+	}
 
 }
